% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_partitions.R
\name{best_partitions}
\alias{best_partitions}
\title{Find Best Partitions}
\usage{
best_partitions(counts, partitions, min_value)
}
\arguments{
\item{counts}{A list of partition counts.}

\item{partitions}{A list of distinct partitions corresponding to the counts.}

\item{min_value}{The minimum count value to be considered.}
}
\value{
A list with two elements: 'count' containing the filtered counts, and 'partitions' containing the corresponding partitions.
}
\description{
This function identifies and returns partitions that have a partition count greater or equal to the specified minimum value.
}
\examples{
counts_list <- list(c(3, 5), c(4, 2))
partitions_list <- list(c("A", "B"), c("C", "D"))
best_partitions(counts_list, partitions_list, min_value = 3)
}
