% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAMP.R
\name{CHAMP}
\alias{CHAMP}
\title{CHAMP function}
\usage{
CHAMP(network, partitions, plottitle = NULL)
}
\arguments{
\item{network}{The network to analyze.}

\item{partitions}{The list of partitions to consider.}

\item{plottitle}{The title of the plot to generate (optional).}
}
\value{
A summary of the CHAMP partitions, including the range of gamma values
for which each partition is part of the CHAMP set and the number of clusters in each partition.
}
\description{
This function identifies the Convex Hull of Admissible Modularity Partitions (CHAMP)
set of partitions and their corresponding domains of optimality.
}
\examples{
# The function takes a network and partitions as input. An example use could be:
# partition_summary <- CHAMP(network, partitions)
}
\author{
Code written in parts by Rachel Matthew, Ryan Rebne, Ava Scharfstein and PJM.
}
