% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_res_parameter.R
\name{derive_res_parameter}
\alias{derive_res_parameter}
\title{Derive Resolution Parameter}
\usage{
derive_res_parameter(partition, network)
}
\arguments{
\item{partition}{A list containing details of the partition. It should include
'nb_clusters' for the number of clusters, and 'membership' for the membership of each element.}

\item{network}{The network graph on which the partition is placed.}
}
\value{
A numeric value representing the resolution parameter, or NULL if there's only one cluster.
}
\description{
This function computes the resolution parameter based on a given partition
and network. The resolution parameter is computed only if there is more
than one cluster in the partition.
}
\examples{
# Assume you have a partition object and a network object defined
derive_res_parameter(partition, network)
}
