% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_max.R
\name{find_max}
\alias{find_max}
\title{Find Maxima in Modularity Matrix}
\usage{
find_max(mod_matrix, gammas)
}
\arguments{
\item{mod_matrix}{A matrix containing 'base' and 'decrement' columns representing the modularity structure.}

\item{gammas}{A numeric vector representing the range of gamma values to consider.}
}
\value{
A list with two elements: 'edges', containing the points of change, and 'partitions', containing the corresponding partitions.
}
\description{
This function identifies points of change and corresponding partitions based on a given
modularity matrix and range of gamma values.
}
\examples{
# Assume mod_matrix is defined with appropriate 'base' and 'decrement' columns
# and gammas is defined as a range of values
find_max(mod_matrix, gammas)
}
