% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CHAMP_Map.R
\name{get_CHAMP_Map}
\alias{get_CHAMP_Map}
\title{Wrapper for Other CHAMP_Map_scripts Helper Functions to Get the Iterative Map}
\usage{
get_CHAMP_Map(network, partitions, partition_summary, plottitle = NULL)
}
\arguments{
\item{network}{A network object representing the structure of the network.}

\item{partitions}{A list containing the different partitions.}

\item{partition_summary}{A data frame containing the summary of the partitions,
obtained by CHAMP.R.}

\item{plottitle}{(Optional) A title for the plot.}
}
\value{
A data frame containing the updated partition_summary with iterative
        map information for all partitions.
}
\description{
This function is a wrapper for other CHAMP_Map_scripts helper functions to
get the iterative map for the resolution parameter on the CHAMP set of
partitions and their corresponding domains of optimality.
}
\examples{
# Example usage of the function (assuming required inputs are defined)
get_CHAMP_Map(network, partitions, partition_summary)

}
\author{
Rachel Matthew, Ryan Rebne, Ava Scharfstein, PJM
}
