% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partitions.R
\name{get_partitions}
\alias{get_partitions}
\title{Get a collection of partitions to run through CHAMP (`get_partitions`)}
\usage{
get_partitions(
  network,
  gamma_range = c(0, 2),
  n_runs = 100,
  n_iterations = 2,
  seed = NULL
)
}
\arguments{
\item{network}{The network to be clustered (community detection).}

\item{gamma_range}{The range of resolution parameters gamma used (default from 0 to 2).}

\item{n_runs}{The number of cluster_leiden() runs to be attempted (default = 100).}

\item{n_iterations}{Parameter to be passed to cluster_leiden (default = 2).}

\item{seed}{Optional random seed for reproducing pseudo-random results.}
}
\value{
A list containing unique partitions generated by `cluster_leiden()`.
Coded in parts by Rachel Matthew, Ryan Rebne, Ava Scharfstein, and PJM.
}
\description{
The `get_partitions` function is a wrapper around other
  CHAMP_Map_scripts helper functions to gather a collection of partitions
  using igraph's cluster_leiden for maximizing modularity with a resolution
  parameter.
}
