% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_gamma.R
\name{iterate_gamma}
\alias{iterate_gamma}
\title{Iterate over a range of gamma values and return the partitions from `cluster_leiden`}
\usage{
iterate_gamma(network)
}
\arguments{
\item{network}{The network to be clustered (community detection).}
}
\value{
A list of partitions generated by `cluster_leiden` over the gamma range.
}
\description{
Calls `cluster_leiden` on the given network for a sequence of gamma
  values, returning the partitions.
}
\examples{
partitions <- iterate_gamma(network)
}
