% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_level_champ.R
\name{partition_level_champ}
\alias{partition_level_champ}
\title{Calculate partition level metrics for CHAMP}
\usage{
partition_level_champ(network, partition)
}
\arguments{
\item{network}{The network to be analyzed.}

\item{partition}{A list representing the partitioning of the network's vertices.}
}
\value{
A list containing the values for `a_partition` and `p_partition`.
}
\description{
Computes the metrics `a_partition` and `p_partition` based on the given network
  and partition.
}
\examples{
result <- partition_level_champ(network, partition)
}
