% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_partitions.R
\name{unique_partitions}
\alias{unique_partitions}
\title{Find Unique Partitions and Their Counts}
\usage{
unique_partitions(partitions)
}
\arguments{
\item{partitions}{A list of partitions.}
}
\value{
A list containing two elements:
  - "count": A list containing the counts of each unique partition.
  - "partitions": A list containing the unique partitions.
}
\description{
Given a list of partitions, finds the unique partitions and the
  number of times each unique partition appeared.
}
\examples{
result <- unique_partitions(partitions)
}
